+++ app/app/src/main/java/com/uu_uce/services/LocationServices.kt
@@ -35,7 +35,7 @@ enum class LocationPollStartResult{
     }
 }
 
-data class UTMCoordinate(val zone : Int, val letter : Char, val east : Double, val north : Double)
+data class UTMCoordinate(val zone : Int, val letter : Char, val east : Float, val north : Float)
 {
     override fun toString(): String {
         return  "$zone" +
@@ -67,7 +67,7 @@ fun degreeToUTM(degPos : p2) : UTMCoordinate{
     val lon = degPos.second
 
     val zone = floor(lon / 6 + 31).toInt()
-    val letter = latToUTMLetter(lat)
+    val letter = latToUTMLetter(lat.toDouble())
 
     val deg = Math.PI / 180
 
@@ -105,7 +105,7 @@ fun degreeToUTM(degPos : p2) : UTMCoordinate{
     if (letter < 'M') northing += 10000000
     northing = (northing * 100).roundToInt() * 0.01
 
-    return UTMCoordinate(zone, letter, easting, northing)
+    return UTMCoordinate(zone, letter, easting.toFloat(), northing.toFloat())
 }
 
 fun latToUTMLetter(lat: Double): Char{
@@ -189,7 +189,7 @@ class LocationServices{
             override fun onLocationChanged(location: Location?) {
                 if (location != null) {
                     lastKnownLocation = location
-                    action(Pair(location.latitude, location.longitude))
+                    action(Pair(location.latitude.toFloat(), location.longitude.toFloat()))
                     Logger.log( LogType.Event,
                         "LocationServices",
                         "Latitude : " + location.latitude
@@ -250,7 +250,7 @@ class LocationServices{
         fun startLocUpdates(provider : String){
             locationManager.requestLocationUpdates(provider, pollTimeMs, minDist, locationListener)
             if(locationManager.getLastKnownLocation(provider) != null)
-                action(Pair(locationManager.getLastKnownLocation(provider).latitude, locationManager.getLastKnownLocation(provider).longitude))
+                action(Pair(locationManager.getLastKnownLocation(provider).latitude.toFloat(), locationManager.getLastKnownLocation(provider).longitude.toFloat()))
             networkRunning = true
             networkProvider = provider
         }
