+++ app/app/src/main/java/com/uu_uce/shapefiles/Camera.kt
@@ -3,12 +3,12 @@ package com.uu_uce.shapefiles
 import kotlin.math.pow
 import kotlin.math.sqrt
 
-fun distXy(p: p3, q: p3): Double{
+fun distXy(p: p3, q: p3): Float{
     return sqrt((p.first - q.first).pow(2) +
             (p.second - q.second).pow(2))
 }
 
-fun lerp(a: Double, b: Double, t: Double): Double{
+fun lerp(a: Float, b: Float, t: Float): Float{
     return a + (b - a) * t
 }
 
@@ -27,23 +27,23 @@ zoom: current zoom/height
 viewMin/viewMax: bounds of the currently loaded layers, which the camera can't leave
  */
 class Camera(
-    private var x: Double,
-    private var y: Double,
-    private var zoom: Double,
+    private var x: Float,
+    private var y: Float,
+    private var zoom: Float,
     private val viewMin: p3,
     private val viewMax: p3){
 
-    private val mx = (viewMin.first + viewMax.first) / 2.0
-    private val my = (viewMin.second + viewMax.second) / 2.0
+    private val mx = (viewMin.first + viewMax.first) / 2.0f
+    private val my = (viewMin.second + viewMax.second) / 2.0f
     private val maxDistXy = distXy(viewMin, viewMax)
     private var velo = p2Zero
 
-    var maxZoom = 1.0
+    var maxZoom = 1.0f
         set(value) {minZoom = value/500; field = value}
-    var minZoom = 0.01
+    var minZoom = 0.01f
 
-    private var lastWoff = 0.0
-    private var lastHoff = 0.0
+    private var lastWoff = 0.0f
+    private var lastHoff = 0.0f
     private var changed = true
 
     private var animType: AnimType = AnimType.NONE
@@ -57,15 +57,15 @@ class Camera(
 
     private var animBegin = p3Zero
     private var animTarget = p3Zero
-    private var animDuration = 0.0
-    private var animStartT = 0.0
-    private var animT = 0.0
+    private var animDuration = 0.0f
+    private var animStartT = 0.0f
+    private var animT = 0.0f
 
     //variables for sliding camera
-    private var decline = p2(1.0,1.0)
-    private var declineLength = 40.0
+    private var decline = p2(1.0f,1.0f)
+    private var declineLength = 40.0f
 
-    var wAspect = 0.0
+    var wAspect = 0.0f
 
     //get matrix for drawing lines
     fun getScaleTrans(): Pair<FloatArray,FloatArray>{
@@ -90,8 +90,8 @@ class Camera(
         val w = viewMax.first - viewMin.first
         val h = viewMax.second - viewMin.second
         val maxwh = maxOf(w,h)
-        val woff = maxwh * wAspect / 2.0 * zoom
-        val hoff = maxwh / 2.0 * zoom
+        val woff = maxwh * wAspect / 2.0f * zoom
+        val hoff = maxwh / 2.0f * zoom
         lastWoff = woff
         lastHoff = hoff
         val nmin = p2(x - woff, y - hoff)
@@ -113,21 +113,21 @@ class Camera(
         changed = true
     }
 
-    private fun setXy(xx: Double, yy: Double){
+    private fun setXy(xx: Float, yy: Float){
         changed = changed || xx != x || yy != y
         x = xx
         y = yy
     }
 
     //set the x and y to new values, while not going out of bounds
-    private fun setPos(newX: Double, newY: Double){
+    private fun setPos(newX: Float, newY: Float){
         if(isBusy() || viewMin.first > viewMax.first || viewMin.second > viewMax.second) return
         val xvalue = newX.coerceIn(viewMin.first,viewMax.first)
         val yvalue = newY.coerceIn(viewMin.second,viewMax.second)
         setXy(xvalue,yvalue)
     }
 
-    fun moveCamera(dx: Double, dy: Double){
+    fun moveCamera(dx: Float, dy: Float){
         if(isBusy()) return
         setPos(x + (dx * lastWoff), y + (dy * lastHoff))
 
@@ -142,50 +142,50 @@ class Camera(
         animType = AnimType.SLIDE
     }
 
-    fun getZoom(): Double{
+    fun getZoom(): Float{
         return zoom
     }
 
-    private fun setZ(zz: Double){
+    private fun setZ(zz: Float){
         changed = changed || zoom != zz
         zoom = zz
     }
 
-    fun setZoom(newZoom: Double){
+    fun setZoom(newZoom: Float){
         if(isBusy()) return
         setZ(newZoom.coerceIn(minZoom, maxZoom))
     }
 
-    fun zoomIn(factor: Double){
+    fun zoomIn(factor: Float){
         if(isBusy()) return
         val z = zoom * factor
         setZ(z.coerceIn(minZoom, maxZoom))
     }
 
     //fully zoom out
-    fun zoomOutMax(duration: Double){
+    fun zoomOutMax(duration: Float){
         if(isBusy()) return
         animBegin = Triple(x, y, zoom)
         animTarget = Triple(mx,my,maxZoom)
-        animStartT = System.currentTimeMillis().toDouble()
+        animStartT = System.currentTimeMillis().toFloat()
         animDuration = duration
         animType = AnimType.OUT
-        animT = 0.0
+        animT = 0.0f
     }
 
     //initialize the animation from current position to target in durationMs milisecs
-    fun startAnimation(target: p3, durationMs: Double){
+    fun startAnimation(target: p3, durationMs: Float){
         if(isBusy()) return
         animBegin = Triple(x, y, zoom)
         animTarget = target
-        animStartT = System.currentTimeMillis().toDouble()
+        animStartT = System.currentTimeMillis().toFloat()
         animDuration = durationMs
         animType = AnimType.TRANS
-        animT = 0.0
+        animT = 0.0f
     }
 
-    private fun smooth(t: Double): Double{
-        return -(t - 1.0).pow(2.0) + 1.0
+    private fun smooth(t: Float): Float{
+        return -(t - 1.0f).pow(2.0f) + 1.0f
     }
 
     //Updates and returns true if viewport has changed
@@ -208,8 +208,8 @@ class Camera(
 
     //animate a full zoomout
     private fun updateOut(){
-        val ct = System.currentTimeMillis().toDouble()
-        val t = ((ct - animStartT) / animDuration).coerceIn(0.0, 1.0)
+        val ct = System.currentTimeMillis().toFloat()
+        val t = ((ct - animStartT) / animDuration).coerceIn(0.0f, 1.0f)
         x = animBegin.first + (animTarget.first - animBegin.first) * t
         y = animBegin.second + (animTarget.second - animBegin.second) * t
         zoom = animBegin.third + (animTarget.third - animBegin.third) * t
@@ -222,12 +222,12 @@ class Camera(
 
     //animate the movement to animTarget
     private fun updateTrans(){
-        val ct = System.currentTimeMillis().toDouble()
-        val t = ((ct - animStartT) / animDuration).coerceIn(0.0, 1.0)
+        val ct = System.currentTimeMillis().toFloat()
+        val t = ((ct - animStartT) / animDuration).coerceIn(0.0f, 1.0f)
         val distFraction = distXy(animBegin, animTarget) / maxDistXy
-        val zoomAvg = (animBegin.third + animTarget.third) / 2.0
+        val zoomAvg = (animBegin.third + animTarget.third) / 2.0f
         val midZoom = (maxZoom - zoomAvg)*distFraction + zoomAvg
-        val zt = 1.0/3.0
+        val zt = 1.0f/3.0f
         when {
             t < zt -> {
                 zoom = lerp(animBegin.third, midZoom, smooth(t / zt))
@@ -254,11 +254,11 @@ class Camera(
         changed = true
         setPos(x + velo.first, y + velo.second)
         val newXVel =
-            if(decline.first > 0) maxOf(0.0,velo.first - decline.first)
-            else minOf(0.0,velo.first - decline.first)
+            if(decline.first > 0) maxOf(0.0f,velo.first - decline.first)
+            else minOf(0.0f,velo.first - decline.first)
         val newYVel =
-            if(decline.second > 0) maxOf(0.0,velo.second - decline.second)
-            else minOf(0.0,velo.second - decline.second)
+            if(decline.second > 0) maxOf(0.0f,velo.second - decline.second)
+            else minOf(0.0f,velo.second - decline.second)
         velo = p2(newXVel,newYVel)
         if(velo == p2Zero) {
             animType = AnimType.NONE
